ammo_trading_agent/requirements.txt
# Core dependencies
streamlit
pandas
numpy
requests
finnhub-python

# For data visualization (optional but recommended)
plotly
matplotlib

# For API key management
python-dotenv